remove absolute refence infavour of barrel file references
